refactor(worker): Simplify dispatcher
